requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==1.10.1
django-cors-headers==2.5.2
djangorestframework==3.4.1
drfdocs==0.0.11
ecdsa==0.13
Fabric==1.12.0
idna==2.9
paramiko==1.18.5
pkg-resources==0.0.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
pycrypto==2.6.1
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.8
